pyproject.toml
[project]
name = "career-sathi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=7.2.7",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "gtts>=2.5.4",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "supabase>=2.27.0",
    "uvicorn>=0.40.0",
]
